+++ user-manager/src/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
 import { LoginComponent } from './pages/login/login.component';
-import { noAuthGuard } from './guards/no-auth-guard';
+import { noAuthGuard } from './auth/guards/no-auth-guard';
 
 export const routes: Routes = [
   {
