+++ user-manager/src/app/auth/directives/password-strength.directive.ts
